Moved files around and cleaned up formatting
