+++ client/src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from "react";
 import { BrowserRouter, Switch, Route } from "react-router-dom";
-import PrivateRoute from "./Components/AppComponents/Auth/PrivateRoute"
-import Nav from "./Components/BootstrapComponents/Nav";
-import Login from "./Components/AppComponents/Auth/Login";
-import Signup from "./Components/AppComponents/Auth/Signup";
-import NoMatch from "./Components/AppComponents/Auth/NoMatch";
-import Profile from "./Components/AppComponents/Profile/ProfileWrapper";
-import Footer from "./Components/BootstrapComponents/Footer";
+import PrivateRoute from "./Components/Auth/PrivateRoute"
+import Nav from "./Components/Layout/Nav";
+import Login from "./Components/Auth/Login";
+import Signup from "./Components/Auth/Signup";
+import NoMatch from "./Components/Auth/NoMatch";
+import Profile from "./Components/Profile/ProfileWrapper";
+import Footer from "./Components/Layout/Footer";
 
 class App extends Component {
   state = {
@@ -48,7 +48,7 @@ class App extends Component {
             <Route component={NoMatch} />
           </Switch>
         </div>
-        <Footer></Footer>
+        <Footer />
       </BrowserRouter>
      
     );
