+++ client/src/Components/Auth/Signup/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Link} from "react-router-dom";
-import API from "../../../../utils/API";
+import API from "../../../utils/API";
 import "./Signup.css";
 
 class Signup extends Component {
