+++ client/src/Components/Profile/CreateChallenges/index.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import API from "../../../../utils/API";
+import API from "../../../utils/API";
 import DatePicker from "react-datepicker";
 import { setMinutes, setHours, addDays } from "date-fns";
 import moment from "moment";
