+++ client/src/Components/Profile/ProfileNav/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Link } from "react-router-dom";
-import API from "../../../../utils/API";
+import API from "../../../utils/API";
 import "./ProfileNav.css";
 
 class ProfileNav extends Component {
