+++ client/src/Components/Profile/ProfileWrapper/ProfileWrapper.css
@@ -34,3 +34,30 @@ ul {
 .nav-link {
   float: left;
 }
+
+.btn-success {
+  background-color: #e74b3c;
+  border-color: #e74b3c;
+  box-shadow: 1px 1px 6px 0px rgba(0, 0, 0, 0.75);
+}
+
+.btn-link {
+  color: #e74b3c !important;
+}
+
+.btn-success:hover {
+  background-color: #df2f1f !important;
+  border-color: #e74b3c !important;
+  color: white !important;
+}
+
+.btn-success:active {
+  background-color: #af3c31 !important;
+  border-color: #e74b3c !important;
+  color: white !important;
+}
+.btn-success:focus {
+  background-color: #af3c31 !important;
+  border-color: #e74b3c !important;
+  color: white !important;
+}
