+++ client/src/Components/Profile/Timeline/index.js
@@ -1,5 +1,4 @@
-import React, { Component } from "react";
-import API from "../../../../utils/API";
+import React from "react";
 import TimelinePost from "../TimelinePost";
 import "./Timeline.css";
